+++ csrc/single_decode_jit_pybind.cu
@@ -18,7 +18,7 @@
 #include "single_decode_config.inc"
 
 void single_decode_with_kv_cache(at::Tensor q, at::Tensor k, at::Tensor v, at::Tensor tmp,
-                                 at::Tensor o, unsigned int layout,
+                                 at::Tensor o, int64_t layout,
                                  int window_left ADDITIONAL_FUNC_PARAMS, int64_t cuda_stream);
 
 TORCH_LIBRARY(TORCH_EXTENSION_NAME, m) {
